+++ protocol/storagepb/delegate_grpc.pb.go
@@ -19,10 +19,11 @@ import (
 const _ = grpc.SupportPackageIsVersion9
 
 const (
-	Delegate_GetDelegates_FullMethodName       = "/storagepb.Delegate/GetDelegates"
-	Delegate_GetDelegateProfile_FullMethodName = "/storagepb.Delegate/GetDelegateProfile"
-	Delegate_GetAllDelegations_FullMethodName  = "/storagepb.Delegate/GetAllDelegations"
-	Delegate_GetAllDelegators_FullMethodName   = "/storagepb.Delegate/GetAllDelegators"
+	Delegate_GetDelegates_FullMethodName        = "/storagepb.Delegate/GetDelegates"
+	Delegate_GetDelegateProfile_FullMethodName  = "/storagepb.Delegate/GetDelegateProfile"
+	Delegate_GetAllDelegations_FullMethodName   = "/storagepb.Delegate/GetAllDelegations"
+	Delegate_GetAllDelegators_FullMethodName    = "/storagepb.Delegate/GetAllDelegators"
+	Delegate_GetDelegatesSummary_FullMethodName = "/storagepb.Delegate/GetDelegatesSummary"
 )
 
 // DelegateClient is the client API for Delegate service.
@@ -36,6 +37,8 @@ type DelegateClient interface {
 	GetAllDelegations(ctx context.Context, in *GetAllDelegationsRequest, opts ...grpc.CallOption) (*GetAllDelegationsResponse, error)
 	// GetAllDelegators returns list of delegators based on internal data grouped by dao
 	GetAllDelegators(ctx context.Context, in *GetAllDelegatorsRequest, opts ...grpc.CallOption) (*GetAllDelegatorsResponse, error)
+	// GetDelegatesSummary returns count of delegators and delegations
+	GetDelegatesSummary(ctx context.Context, in *GetDelegatesSummaryRequest, opts ...grpc.CallOption) (*GetDelegatesSummaryResponse, error)
 }
 
 type delegateClient struct {
@@ -86,6 +89,16 @@ func (c *delegateClient) GetAllDelegators(ctx context.Context, in *GetAllDelegat
 	return out, nil
 }
 
+func (c *delegateClient) GetDelegatesSummary(ctx context.Context, in *GetDelegatesSummaryRequest, opts ...grpc.CallOption) (*GetDelegatesSummaryResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(GetDelegatesSummaryResponse)
+	err := c.cc.Invoke(ctx, Delegate_GetDelegatesSummary_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // DelegateServer is the server API for Delegate service.
 // All implementations must embed UnimplementedDelegateServer
 // for forward compatibility.
@@ -97,6 +110,8 @@ type DelegateServer interface {
 	GetAllDelegations(context.Context, *GetAllDelegationsRequest) (*GetAllDelegationsResponse, error)
 	// GetAllDelegators returns list of delegators based on internal data grouped by dao
 	GetAllDelegators(context.Context, *GetAllDelegatorsRequest) (*GetAllDelegatorsResponse, error)
+	// GetDelegatesSummary returns count of delegators and delegations
+	GetDelegatesSummary(context.Context, *GetDelegatesSummaryRequest) (*GetDelegatesSummaryResponse, error)
 	mustEmbedUnimplementedDelegateServer()
 }
 
@@ -118,6 +133,9 @@ func (UnimplementedDelegateServer) GetAllDelegations(context.Context, *GetAllDel
 }
 func (UnimplementedDelegateServer) GetAllDelegators(context.Context, *GetAllDelegatorsRequest) (*GetAllDelegatorsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAllDelegators not implemented")
+}
+func (UnimplementedDelegateServer) GetDelegatesSummary(context.Context, *GetDelegatesSummaryRequest) (*GetDelegatesSummaryResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetDelegatesSummary not implemented")
 }
 func (UnimplementedDelegateServer) mustEmbedUnimplementedDelegateServer() {}
 func (UnimplementedDelegateServer) testEmbeddedByValue()                  {}
@@ -212,6 +230,24 @@ func _Delegate_GetAllDelegators_Handler(srv interface{}, ctx context.Context, de
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Delegate_GetDelegatesSummary_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetDelegatesSummaryRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DelegateServer).GetDelegatesSummary(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Delegate_GetDelegatesSummary_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DelegateServer).GetDelegatesSummary(ctx, req.(*GetDelegatesSummaryRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // Delegate_ServiceDesc is the grpc.ServiceDesc for Delegate service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -234,6 +270,10 @@ var Delegate_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GetAllDelegators",
 			Handler:    _Delegate_GetAllDelegators_Handler,
+		},
+		{
+			MethodName: "GetDelegatesSummary",
+			Handler:    _Delegate_GetDelegatesSummary_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},