+++ CHANGELOG.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Implement getting delegations by address
+- Implement getting delegators by address
+- Implement getting count of delegations by address
 
 ## [0.3.1] - 2024-10-18